Update winblows-install.md
Hey, large FAT is not available in the newest Version. I tried exFAT and failed.
FAT32 should be the right setting, right? My change propose would be to update the rufus image with the correct settings.
Of course I can't upload files. Can you do it for me?

Thanks in advance! :)